+++ app/model.go
@@ -5,7 +5,8 @@ import (
 
 	"github.com/charmbracelet/bubbles/list"
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/masamerc/sevp/src/internal"
+
+	"github.com/masamerc/sevp/internal"
 )
 
 type Model struct {
