+++ cmd/view.go
@@ -7,10 +7,11 @@ import (
 	"fmt"
 
 	"github.com/charmbracelet/lipgloss"
-	"github.com/masamerc/sevp/src/app"
-	"github.com/masamerc/sevp/src/internal"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
+
+	"github.com/masamerc/sevp/app"
+	"github.com/masamerc/sevp/internal"
 )
 
 // view command will display the details of a selector.
