+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/masamerc/sevp/src/cmd"
+import "github.com/masamerc/sevp/cmd"
 
 func main() {
 	cmd.Execute()
